database structure changed
database structure changed for permissions

+++ app/database/models/index.js
@@ -1,11 +1,21 @@
 const colors = require('colors');
-
+const { connectDB } = require('../../../config/database');
+const mongoose = require('mongoose');
 const fs = require('fs');
 const path = require('path');
 const basename = path.basename(__filename);
 const models = [];
 
 const migration = async () => {
+  if (process.env.NODE_ENV === "development" && process.argv.includes('db')) {
+    await mongoose.connection.db.dropDatabase(function (err, result) {
+      if (result) {
+        console.log(colors.red(`database deleted`));
+      } else {
+        console.log(err);
+      }
+    });
+  }
   await fs
     .readdirSync(__dirname)
     .filter(file => {
@@ -15,7 +25,7 @@ const migration = async () => {
       const model = require(path.join(__dirname, file));
       models.push(model);
     });
-  if (process.env.NODE_ENV === "development") {
+  if (process.env.NODE_ENV === "development" && process.argv.includes('data')) {
     models.map(async (value) => {
       await value.deleteMany({});
     });
